third_party/WebKit/LayoutTests/accessibility/aria-tree.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<ul id="manual-props" role="tree">
  <li role="none">
    <span id="manual-item" role="treeitem"
          aria-level="2" aria-setsize="5" aria-posinset="3">
      Oranges
    </span>
  </li>
</ul>

<ul id="auto-props" role="tree">
  <li role="treeitem">
    Veggies
    <ul role="group">
      <li role="treeitem">
          Green beans
      </li>
      <li role="treeitem">
        Onions
        <ul role="group">
          <li role="treeitem">
            Green
          </li>
          <li role="treeitem" aria-hidden="true">
            Hidden
          </li>
          <li id="auto-item" role="treeitem">
            Vidalia
          </li>
          <li role="treeitem">
            Yellow
          </li>
        </ul>
      </li>
    </ul>
  </li>
</ul>

<ul id="auto-set-props-flat" role="tree">
  <li role="treeitem" aria-level="1">
    Cats
  </li>
  <li role="treeitem" aria-level="2">
    Hypoallegenic
  </li>
  <li role="treeitem" aria-level="3">
    Siberian
  </li>
  <li role="treeitem" aria-level="3">
    Sphynx
  </li>
  <li role="treeitem" aria-level="2">
    Siamese
  </li>
  <li role="treeitem" id="flat-auto-item" aria-level="2">
    Calico
  </li>
  <li role="treeitem" aria-level="1">
    Fish
  </li>
</ul>

<script>
function axElementById(id) {
    return accessibilityController.accessibleElementById(id);
}

test(function(t) {
    var manualItem = axElementById("manual-item");
    assert_equals(manualItem.hierarchicalLevel, 2);
}, "A manual tree item exposes its level property");

test(function(t) {
    var manualItem = axElementById("manual-item");
    assert_equals(manualItem.posInSet, 3);
}, "A manual tree item exposes its posinset property");

test(function(t) {
    var manualItem = axElementById("manual-item");
    assert_equals(manualItem.setSize, 5);
}, "A manual tree item exposes its setsize property");

test(function(t) {
    var autoItem = axElementById("auto-item");
    assert_equals(autoItem.hierarchicalLevel, 3);
}, "An auto tree item computes its level property");

test(function(t) {
    var autoItem = axElementById("auto-item");
    assert_equals(autoItem.posInSet, 2);
}, "An auto tree item computes its posinset property");

test(function(t) {
    var autoItem = axElementById("auto-item");
    assert_equals(autoItem.setSize, 3);
}, "An auto tree item computes its setsize property");

test(function(t) {
    var flatAutoItem = axElementById("flat-auto-item");
    assert_equals(flatAutoItem.setSize, 3);
}, "A flatauto tree item with levels computes its setsize property");

test(function(t) {
    var flatAutoItem = axElementById("flat-auto-item");
    assert_equals(flatAutoItem.posInSet, 3);
}, "A flat auto tree item with levels computes its posinset property");

</script>
